Move tuple into its own file
